Kill index rebuilding, merging code for LD_1
