+++ go/libraries/doltcore/table/editor/index_editor_test.go
@@ -399,333 +399,6 @@ func TestIndexEditorUniqueErrorDoesntPersist(t *testing.T) {
 	require.NoError(t, indexEditor.InsertRow(context.Background(), fullKey, partialKey, value))
 }
 
-func TestIndexRebuildingWithZeroIndexes(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	tSchema := createTestSchema(t)
-	_, err := tSchema.Indexes().RemoveIndex(testSchemaIndexName)
-	require.NoError(t, err)
-	_, err = tSchema.Indexes().RemoveIndex(testSchemaIndexAge)
-	require.NoError(t, err)
-	rowData, _ := createTestRowData(t, vrw, tSchema)
-
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, tSchema, rowData)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	rebuildAllTable, err := RebuildAllIndexes(context.Background(), originalTable, opts)
-	require.NoError(t, err)
-	_, err = rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexName)
-	require.Error(t, err)
-
-	_, err = RebuildIndex(context.Background(), originalTable, testSchemaIndexName, opts)
-	require.Error(t, err)
-}
-
-func TestIndexRebuildingWithOneIndex(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	tSchema := createTestSchema(t)
-	_, err := tSchema.Indexes().RemoveIndex(testSchemaIndexAge)
-	require.NoError(t, err)
-	index := tSchema.Indexes().GetByName(testSchemaIndexName)
-	require.NotNil(t, index)
-	indexSch := index.Schema()
-	rowData, rows := createTestRowData(t, vrw, tSchema)
-
-	indexExpectedRows := make([]row.Row, len(rows))
-	for i, r := range rows {
-		indexKey := make(row.TaggedValues)
-		for _, tag := range index.AllTags() {
-			val, ok := r.GetColVal(tag)
-			require.True(t, ok)
-			indexKey[tag] = val
-		}
-		indexExpectedRows[i], err = row.New(types.Format_LD_1, indexSch, indexKey)
-		require.NoError(t, err)
-	}
-
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, tSchema, rowData)
-	require.NoError(t, err)
-
-	var indexRows []row.Row
-
-	opts := TestEditorOptions(vrw)
-	rebuildAllTable, err := RebuildAllIndexes(context.Background(), originalTable, opts)
-	require.NoError(t, err)
-	indexRowData, err := rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexName)
-	require.NoError(t, err)
-	_ = indexRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexExpectedRows, indexRows)
-
-	indexRows = nil
-	indexRowData, err = RebuildIndex(context.Background(), originalTable, testSchemaIndexName, opts)
-	require.NoError(t, err)
-	_ = indexRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexExpectedRows, indexRows)
-}
-
-func TestIndexRebuildingWithTwoIndexes(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	tSchema := createTestSchema(t)
-
-	indexName := tSchema.Indexes().GetByName(testSchemaIndexName)
-	require.NotNil(t, indexName)
-	indexAge := tSchema.Indexes().GetByName(testSchemaIndexAge)
-	require.NotNil(t, indexAge)
-
-	indexNameSch := indexName.Schema()
-	indexAgeSch := indexAge.Schema()
-
-	rowData, rows := createTestRowData(t, vrw, tSchema)
-	indexNameExpectedRows, indexAgeExpectedRows := rowsToIndexRows(t, rows, indexName, indexAge)
-
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, tSchema, rowData)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	rebuildAllTable := originalTable
-	var indexRows []row.Row
-
-	// do two runs, data should not be different regardless of how many times it's ran
-	for i := 0; i < 2; i++ {
-		rebuildAllTable, err = RebuildAllIndexes(context.Background(), rebuildAllTable, opts)
-		require.NoError(t, err)
-
-		indexNameRowData, err := rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexName)
-		require.NoError(t, err)
-		_ = indexNameRowData.IterAll(context.Background(), func(key, value types.Value) error {
-			indexRow, err := row.FromNoms(indexNameSch, key.(types.Tuple), value.(types.Tuple))
-			require.NoError(t, err)
-			indexRows = append(indexRows, indexRow)
-			return nil
-		})
-		assert.ElementsMatch(t, indexNameExpectedRows, indexRows)
-		indexRows = nil
-
-		indexAgeRowData, err := rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexAge)
-		require.NoError(t, err)
-		_ = indexAgeRowData.IterAll(context.Background(), func(key, value types.Value) error {
-			indexRow, err := row.FromNoms(indexAgeSch, key.(types.Tuple), value.(types.Tuple))
-			require.NoError(t, err)
-			indexRows = append(indexRows, indexRow)
-			return nil
-		})
-		assert.ElementsMatch(t, indexAgeExpectedRows, indexRows)
-		indexRows = nil
-
-		indexNameRowData, err = RebuildIndex(context.Background(), originalTable, testSchemaIndexName, opts)
-		require.NoError(t, err)
-		_ = indexNameRowData.IterAll(context.Background(), func(key, value types.Value) error {
-			indexRow, err := row.FromNoms(indexNameSch, key.(types.Tuple), value.(types.Tuple))
-			require.NoError(t, err)
-			indexRows = append(indexRows, indexRow)
-			return nil
-		})
-		assert.ElementsMatch(t, indexNameExpectedRows, indexRows)
-		indexRows = nil
-
-		indexAgeRowData, err = RebuildIndex(context.Background(), originalTable, testSchemaIndexAge, opts)
-		require.NoError(t, err)
-		_ = indexAgeRowData.IterAll(context.Background(), func(key, value types.Value) error {
-			indexRow, err := row.FromNoms(indexAgeSch, key.(types.Tuple), value.(types.Tuple))
-			require.NoError(t, err)
-			indexRows = append(indexRows, indexRow)
-			return nil
-		})
-		assert.ElementsMatch(t, indexAgeExpectedRows, indexRows)
-		indexRows = nil
-	}
-
-	// change the underlying data and verify that rebuild changes the data as well
-	rowData, rows = createUpdatedTestRowData(t, vrw, tSchema)
-	indexNameExpectedRows, indexAgeExpectedRows = rowsToIndexRows(t, rows, indexName, indexAge)
-	updatedTable, err := rebuildAllTable.UpdateNomsRows(context.Background(), rowData)
-	require.NoError(t, err)
-	rebuildAllTable, err = RebuildAllIndexes(context.Background(), updatedTable, opts)
-	require.NoError(t, err)
-
-	indexNameRowData, err := rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexName)
-	require.NoError(t, err)
-	_ = indexNameRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexNameSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexNameExpectedRows, indexRows)
-	indexRows = nil
-
-	indexAgeRowData, err := rebuildAllTable.GetNomsIndexRowData(context.Background(), testSchemaIndexAge)
-	require.NoError(t, err)
-	_ = indexAgeRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexAgeSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexAgeExpectedRows, indexRows)
-	indexRows = nil
-
-	indexNameRowData, err = RebuildIndex(context.Background(), updatedTable, testSchemaIndexName, opts)
-	require.NoError(t, err)
-	_ = indexNameRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexNameSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexNameExpectedRows, indexRows)
-	indexRows = nil
-
-	indexAgeRowData, err = RebuildIndex(context.Background(), updatedTable, testSchemaIndexAge, opts)
-	require.NoError(t, err)
-	_ = indexAgeRowData.IterAll(context.Background(), func(key, value types.Value) error {
-		indexRow, err := row.FromNoms(indexAgeSch, key.(types.Tuple), value.(types.Tuple))
-		require.NoError(t, err)
-		indexRows = append(indexRows, indexRow)
-		return nil
-	})
-	assert.ElementsMatch(t, indexAgeExpectedRows, indexRows)
-}
-
-func TestIndexRebuildingUniqueSuccessOneCol(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	colColl := schema.NewColCollection(
-		schema.NewColumn("pk1", 1, types.IntKind, true, schema.NotNullConstraint{}),
-		schema.NewColumn("v1", 2, types.IntKind, false),
-		schema.NewColumn("v2", 3, types.IntKind, false),
-	)
-	sch, err := schema.SchemaFromCols(colColl)
-	require.NoError(t, err)
-	rowData, _ := createTestRowDataFromTaggedValues(t, vrw, sch,
-		row.TaggedValues{1: types.Int(1), 2: types.Int(1), 3: types.Int(1)},
-		row.TaggedValues{1: types.Int(2), 2: types.Int(2), 3: types.Int(2)},
-		row.TaggedValues{1: types.Int(3), 2: types.Int(3), 3: types.Int(3)},
-	)
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, sch, rowData)
-	require.NoError(t, err)
-
-	index, err := sch.Indexes().AddIndexByColTags("idx_v1", []uint64{2}, nil, schema.IndexProperties{IsUnique: true, Comment: ""})
-	require.NoError(t, err)
-	updatedTable, err := originalTable.UpdateSchema(context.Background(), sch)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	_, err = RebuildAllIndexes(context.Background(), updatedTable, opts)
-	require.NoError(t, err)
-
-	_, err = RebuildIndex(context.Background(), updatedTable, index.Name(), opts)
-	require.NoError(t, err)
-}
-
-func TestIndexRebuildingUniqueSuccessTwoCol(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	colColl := schema.NewColCollection(
-		schema.NewColumn("pk1", 1, types.IntKind, true, schema.NotNullConstraint{}),
-		schema.NewColumn("v1", 2, types.IntKind, false),
-		schema.NewColumn("v2", 3, types.IntKind, false),
-	)
-	sch, err := schema.SchemaFromCols(colColl)
-	require.NoError(t, err)
-	rowData, _ := createTestRowDataFromTaggedValues(t, vrw, sch,
-		row.TaggedValues{1: types.Int(1), 2: types.Int(1), 3: types.Int(1)},
-		row.TaggedValues{1: types.Int(2), 2: types.Int(1), 3: types.Int(2)},
-		row.TaggedValues{1: types.Int(3), 2: types.Int(2), 3: types.Int(2)},
-	)
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, sch, rowData)
-	require.NoError(t, err)
-
-	index, err := sch.Indexes().AddIndexByColTags("idx_v1", []uint64{2, 3}, nil, schema.IndexProperties{IsUnique: true, Comment: ""})
-	require.NoError(t, err)
-	updatedTable, err := originalTable.UpdateSchema(context.Background(), sch)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	_, err = RebuildAllIndexes(context.Background(), updatedTable, opts)
-	require.NoError(t, err)
-
-	_, err = RebuildIndex(context.Background(), updatedTable, index.Name(), opts)
-	require.NoError(t, err)
-}
-
-func TestIndexRebuildingUniqueFailOneCol(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	colColl := schema.NewColCollection(
-		schema.NewColumn("pk1", 1, types.IntKind, true, schema.NotNullConstraint{}),
-		schema.NewColumn("v1", 2, types.IntKind, false),
-		schema.NewColumn("v2", 3, types.IntKind, false),
-	)
-	sch, err := schema.SchemaFromCols(colColl)
-	require.NoError(t, err)
-	rowData, _ := createTestRowDataFromTaggedValues(t, vrw, sch,
-		row.TaggedValues{1: types.Int(1), 2: types.Int(1), 3: types.Int(1)},
-		row.TaggedValues{1: types.Int(2), 2: types.Int(2), 3: types.Int(2)},
-		row.TaggedValues{1: types.Int(3), 2: types.Int(2), 3: types.Int(3)},
-	)
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, sch, rowData)
-	require.NoError(t, err)
-
-	index, err := sch.Indexes().AddIndexByColTags("idx_v1", []uint64{2}, nil, schema.IndexProperties{IsUnique: true, Comment: ""})
-	require.NoError(t, err)
-	updatedTable, err := originalTable.UpdateSchema(context.Background(), sch)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	_, err = RebuildAllIndexes(context.Background(), updatedTable, opts)
-	require.Error(t, err)
-
-	_, err = RebuildIndex(context.Background(), updatedTable, index.Name(), opts)
-	require.Error(t, err)
-}
-
-func TestIndexRebuildingUniqueFailTwoCol(t *testing.T) {
-	_, vrw, ns, _ := dbfactory.MemFactory{}.CreateDB(context.Background(), types.Format_LD_1, nil, nil)
-	require.Equal(t, types.Format_LD_1, vrw.Format())
-	colColl := schema.NewColCollection(
-		schema.NewColumn("pk1", 1, types.IntKind, true, schema.NotNullConstraint{}),
-		schema.NewColumn("v1", 2, types.IntKind, false),
-		schema.NewColumn("v2", 3, types.IntKind, false),
-	)
-	sch, err := schema.SchemaFromCols(colColl)
-	require.NoError(t, err)
-	rowData, _ := createTestRowDataFromTaggedValues(t, vrw, sch,
-		row.TaggedValues{1: types.Int(1), 2: types.Int(1), 3: types.Int(1)},
-		row.TaggedValues{1: types.Int(2), 2: types.Int(1), 3: types.Int(2)},
-		row.TaggedValues{1: types.Int(3), 2: types.Int(2), 3: types.Int(2)},
-		row.TaggedValues{1: types.Int(4), 2: types.Int(1), 3: types.Int(2)},
-	)
-	originalTable, err := createTableWithoutIndexRebuilding(context.Background(), vrw, ns, sch, rowData)
-	require.NoError(t, err)
-
-	index, err := sch.Indexes().AddIndexByColTags("idx_v1", []uint64{2, 3}, nil, schema.IndexProperties{IsUnique: true, Comment: ""})
-	require.NoError(t, err)
-	updatedTable, err := originalTable.UpdateSchema(context.Background(), sch)
-	require.NoError(t, err)
-
-	opts := TestEditorOptions(vrw)
-	_, err = RebuildAllIndexes(context.Background(), updatedTable, opts)
-
-	require.Error(t, err)
-
-	_, err = RebuildIndex(context.Background(), updatedTable, index.Name(), opts)
-	require.Error(t, err)
-}
-
 func TestIndexEditorCapacityExceeded(t *testing.T) {
 	// In the event that we reach the iea capacity on Undo, we need to verify that all code paths fail and remain failing
 	ctx := context.Background()
